requirements.txt
uvicorn[standard]
fastapi==0.111.0
pydantic[Email]
python-dotenv==1.0.1
pymongo==4.7.2
passlib==1.7.4
bcrypt==4.0.0
python-jose[cryptography]==3.3.0
python-multipart==0.0.9
httpx==0.27.0
pytest==8.2.0
fastapi-limiter==0.1.6
beanie==1.28.0
pillow==11.1.0
APScheduler==3.11.0
pytz==2024.2